Build template JS in Gulp
Import existing govuk_template JS, and change the Sprockets dependency tree to use Rollup. This builds to the dist folder via new Gulp tasks.
