+++ app/assets/js/template/govuk-template-ie.js
@@ -1,0 +1,2 @@
+import './vendor/html5shiv-printshiv.js'
+import './vendor/json2.js'
