+++ app/assets/js/template/govuk-template.js
@@ -1,0 +1,3 @@
+import './cookie-functions.js'
+import './cookie-bar.js'
+import './core.js'
