+++ app/templates/unbranded_template.html
@@ -14,7 +14,7 @@
 
     <!--[if IE 8]><link href="{{ asset_path + 'stylesheets/fonts-ie8.css' }}" media="all" rel="stylesheet" /><![endif]-->
     <!--[if gte IE 9]><!--><link href="{{ asset_path + 'stylesheets/fonts.css' }}" media="all" rel="stylesheet" /><!--<![endif]-->
-    <!--[if lt IE 9]><script src="{{ asset_path + 'javascripts/ie.js' }}"></script><![endif]-->
+    <!--[if lt IE 9]><script src="{{ asset_path + 'javascripts/govuk-template-ie.js' }}"></script><![endif]-->
 
     <link rel="shortcut icon" href="{{ asset_path + 'images/favicon.ico' }}" type="image/x-icon" />
     <link rel="apple-touch-icon-precomposed" sizes="152x152" href="{{ asset_path + 'images/apple-touch-icon-152x152.png' }}">
