+++ config/paths.js
@@ -2,9 +2,11 @@ const paths = {
   config: 'config/',
   assets: 'app/assets/',
   assetsScss: 'app/assets/scss/',
+  assetsJs: 'app/assets/js/',
   dist: 'dist/',
   distCSS: 'dist/assets/css/',
   distScss: 'dist/assets/scss/',
+  distJs: 'dist/assets/javascripts/',
   templates: 'app/templates/',
   npm: 'node_modules/',
   specs: 'test/specs/'
